+++ packages/lib/gpcircuits/package.json
@@ -28,10 +28,12 @@
     "prepublishOnly": "yarn clean && yarn build",
     "test": "ts-mocha --config ../../../.mocharc.js --exit 'test/**/*.spec.ts'",
     "clean": "rm -rf dist node_modules *.tsbuildinfo build circuits/test ptau",
+    "gen-circuit-parameters": "npx tsx scripts/gen-circuit-parameters.ts",
     "gen-test-artifacts": "npx tsx scripts/gen-test-artifacts.ts"
   },
   "dependencies": {
     "@pcd/pod": "0.1.0",
+    "@types/lodash": "^4.17.0",
     "snarkjs": "^0.7.3"
   },
   "devDependencies": {
