+++ packages/lib/gpcircuits/tsconfig.esm.json
@@ -6,7 +6,7 @@
     "tsBuildInfoFile": "./tsconfig.esm.tsbuildinfo"
   },
   // For CJS/ESM builds, do not include test files
-  "include": ["src", "artifacts/*.json"],
+  "include": ["src", "src/*.json", "artifacts/*.json"],
   // DO NOT MODIFY MANUALLY BEYOND THIS POINT
   // References are automatically maintained by `yarn fix-references`
   "references": [
